refactor(shared-server): remove group-state singleton subtrees
