+++ packages/shared-server/rallar-system/group-state/mutation/idempotency/group-mutation-idempotency-key.ts
@@ -1,4 +1,4 @@
-import type { GroupMutationCommand } from './group-mutation-contracts.ts';
+import type { GroupMutationCommand } from '../group-mutation-contracts.ts';
 
 export function groupMutationIdempotencyKey(command: GroupMutationCommand): string | null {
     return command.requestId === null ? null : command.commandId;
