+++ packages/tests/shared-server/rallar-system/state-events/group-state-event-row-codec.test.ts
